services/ingestion-api/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
click==8.3.2
fastapi==0.115.0
h11==0.16.0
idna==3.11
kafka-python-ng==2.2.3
prometheus_client==0.21.0
psycopg2-binary==2.9.11
pydantic==2.9.0
pydantic_core==2.23.2
python-dotenv==1.0.1
starlette==0.38.6
typing_extensions==4.15.0
tzdata==2026.1
uvicorn==0.32.0
